@@ -1,12 +1,13 @@
 {
   "name": "dylanbonneau",
-  "version": "17.0.0",
+  "version": "1.0.0",
   "scripts": {
     "ng": "ng",
+    "serve:ssr:dylanBonneau": "node dist/dylanbonneau/server/server.mjs",
     "start": "ng serve",
     "start:ci": "ng serve --configuration production",
     "build": "ng build",
-    "build:ci": "ng build --configuration production --build-optimizer --optimization",
+    "build:ci": "ng build --configuration production --optimization",
     "watch": "ng build --watch --configuration development",
     "watch:ci": "ng build --watch --configuration production",
     "test": "ng test",
@@ -20,47 +21,64 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.2",
-    "@angular/cdk": "^17.1.2",
-    "@angular/common": "^17.1.2",
-    "@angular/compiler": "^17.1.2",
-    "@angular/core": "^17.1.2",
-    "@angular/forms": "^17.1.2",
-    "@angular/material": "^17.1.2",
-    "@angular/platform-browser": "^17.1.2",
-    "@angular/platform-browser-dynamic": "^17.1.2",
-    "@angular/router": "^17.1.2",
-    "@commitlint/cli": "^18.6.0",
+    "@angular/animations": "^17.3.0",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/forms": "^17.3.0",
+    "@angular/material": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/platform-server": "^17.3.0",
+    "@angular/router": "^17.3.0",
+    "@angular/ssr": "^17.3.7",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
-    "commitizen": "^4.2.2",
+    "express": "^4.18.2",
+    "rxjs": "~7.8.0",
+    "@commitlint/cli": "^18.6.0",
     "cz-customizable": "^6.3.0",
-    "rxjs": "^7.8.1",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.3",
     "stylelint": "^16.2.1",
     "stylelint-config-property-sort-order-smacss": "^10.0.0",
     "stylelint-declaration-block-no-ignored-properties": "^2.8.0",
     "stylelint-no-unsupported-browser-features": "^8.0.1",
-    "stylelint-scss": "^6.1.0",
-    "tslib": "^2.3.0",
-    "tslint": "^6.1.3",
-    "zone.js": "~0.14.3"
+    "stylelint-scss": "^6.1.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.2",
-    "@angular/cli": "^17.1.2",
-    "@angular/compiler-cli": "^17.1.2",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
+    "@angular-devkit/build-angular": "^17.3.7",
+    "@angular-eslint/builder": "17.4.1",
+    "@angular-eslint/eslint-plugin": "17.4.1",
+    "@angular-eslint/eslint-plugin-template": "17.4.1",
+    "@angular-eslint/schematics": "17.4.1",
+    "@angular-eslint/template-parser": "17.4.1",
+    "@angular/cli": "^17.3.7",
+    "@angular/compiler-cli": "^17.3.0",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/node": "^18.18.0",
+    "@typescript-eslint/eslint-plugin": "7.8.0",
+    "@typescript-eslint/parser": "7.8.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-complex-logic": "^0.3.1",
+    "eslint-plugin-deprecation": "^2.0.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-max-return-statements-per-function": "^0.2.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-sonarjs": "^1.0.3",
+    "eslint-plugin-unicorn": "^53.0.0",
     "highlight.js": "^11.9.0",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
     "ngx-highlightjs": "^10.0.0",
-    "source-map-explorer": "^2.5.3",
-    "typescript": "^5.2"
+    "prettier": "^3.2.5",
+    "typescript": "~5.4.2"
   },
   "config": {
     "commitizen": {
