@@ -1,6 +1,6 @@
 const czConfig = require('./.cz-config.js');
 
-const typeEnum = czConfig.types.map(type => type.value);
+const typeEnum = czConfig.types.map((type) => type.value);
 
 module.exports = {
   rules: {
@@ -8,11 +8,7 @@ module.exports = {
     'footer-leading-blank': [1, 'always'],
     'scope-case': [2, 'always', 'lower-case'],
     'scope-empty': [2, 'never'],
-    'subject-case': [
-      2,
-      'never',
-      ['sentence-case', 'start-case', 'pascal-case', 'upper-case']
-    ],
+    'subject-case': [2, 'never', ['sentence-case', 'start-case', 'pascal-case', 'upper-case']],
     'subject-empty': [2, 'never'],
     'subject-full-stop': [2, 'never', '.'],
     'type-case': [2, 'always', 'lower-case'],
