@@ -1,14 +1,14 @@
 export class Gallery {
-  createdTime: string;
-  fields: {
+  public createdTime: string;
+  public fields: {
     place: string;
     country: string;
     couple: string;
     weddingDate: string;
     text: string;
     card: any;
-  }
-  id: string;
+  };
+  public id: string;
 
   constructor(json?: Gallery) {
     this.createdTime = json?.createdTime ?? '';
@@ -19,7 +19,7 @@ export class Gallery {
       weddingDate: json?.fields.weddingDate ?? '',
       text: json?.fields.text ?? '',
       card: json?.fields.card ?? null
-    }
+    };
     this.id = json?.id ?? '';
   }
 }
