@@ -1,30 +1,33 @@
+import { DOCUMENT } from '@angular/common';
 import { Injectable, Inject } from '@angular/core';
 import { Meta, Title } from '@angular/platform-browser';
-import { DOCUMENT } from '@angular/common';
 
 @Injectable({
-    providedIn: 'root'
+  providedIn: 'root'
 })
 export class SEOService {
-    constructor(@Inject(DOCUMENT) private dom: any, private title: Title, private meta: Meta) { }
+  constructor(
+    @Inject(DOCUMENT) private readonly document: Document,
+    private title: Title,
+    private meta: Meta
+  ) {}
 
+  public updateTitle(title: string): void {
+    this.title.setTitle(title);
+  }
 
-    updateTitle(title: string) {
-        this.title.setTitle(title);
+  public updateDescription(desc: string): void {
+    this.meta.updateTag({ name: 'description', content: desc });
+  }
+
+  public createLinkForCanonicalURL(url?: string): void {
+    const head = this.document.getElementsByTagName('head')[0];
+    let element: HTMLLinkElement | null = this.document.querySelector(`link[rel='canonical']`) || null;
+    if (element === null) {
+      element = this.document.createElement('link');
+      head.appendChild(element);
     }
-
-    updateDescription(desc: string) {
-        this.meta.updateTag({ name: 'description', content: desc })
-    }
-
-    createLinkForCanonicalURL(url?: string) {
-        const head = this.dom.getElementsByTagName('head')[0];
-        var element: HTMLLinkElement = this.dom.querySelector(`link[rel='canonical']`) || null
-        if (element == null) {
-            element = this.dom.createElement('link') as HTMLLinkElement;
-            head.appendChild(element);
-        }
-        element.setAttribute('rel', 'canonical')
-        element.setAttribute('href', url ?? this.dom.URL)
-    }
+    element.setAttribute('rel', 'canonical');
+    element.setAttribute('href', url ?? this.document.URL);
+  }
 }
