@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ButtonsLibraryComponent } from './buttons-library.component';
 
@@ -6,12 +6,11 @@ describe('ButtonsLibraryComponent', () => {
   let component: ButtonsLibraryComponent;
   let fixture: ComponentFixture<ButtonsLibraryComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ButtonsLibraryComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ButtonsLibraryComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ButtonsLibraryComponent);
