@@ -1,10 +1,13 @@
+
 import { Component } from '@angular/core';
+import { MatTabsModule } from '@angular/material/tabs';
+
+import { ButtonsLibraryComponent } from './buttons-library/buttons-library.component';
 
 @Component({
   selector: 'app-buttons',
-  templateUrl: './buttons.component.html',
-  styleUrls: ['./buttons.component.scss']
+  standalone: true,
+  imports: [MatTabsModule, ButtonsLibraryComponent],
+  templateUrl: './buttons.component.html'
 })
-export class ButtonsComponent {
-
-}
+export class ButtonsComponent {}
