@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ColorsLibraryComponent } from './colors-library.component';
 
@@ -6,12 +6,11 @@ describe('ColorsLibraryComponent', () => {
   let component: ColorsLibraryComponent;
   let fixture: ComponentFixture<ColorsLibraryComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ColorsLibraryComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ColorsLibraryComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ColorsLibraryComponent);
