@@ -1,15 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+
+import { Component } from '@angular/core';
+import { MatTabsModule } from '@angular/material/tabs';
+
+import { FormsLibraryComponent } from './forms-library/forms-library.component';
 
 @Component({
   selector: 'app-forms',
-  templateUrl: './forms.component.html',
-  styleUrls: ['./forms.component.scss']
+  standalone: true,
+  imports: [MatTabsModule, FormsLibraryComponent],
+  templateUrl: './forms.component.html'
 })
-export class FormsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
+export class FormsComponent {}
