@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { IconsContributeComponent } from './icons-contribute.component';
 
@@ -6,12 +6,11 @@ describe('IconsContributeComponent', () => {
   let component: IconsContributeComponent;
   let fixture: ComponentFixture<IconsContributeComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ IconsContributeComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [IconsContributeComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(IconsContributeComponent);
