@@ -1,18 +1,16 @@
-import { Component, OnInit } from '@angular/core';
+
+import { Component } from '@angular/core';
+import { HighlightModule } from 'ngx-highlightjs';
 
 @Component({
   selector: 'app-icons-contribute',
+  standalone: true,
+  imports: [HighlightModule],
   templateUrl: './icons-contribute.component.html',
   styleUrls: ['./icons-contribute.component.scss']
 })
-export class IconsContributeComponent implements OnInit {
+export class IconsContributeComponent {
   public newIcon = `.icon-new-icon:before {
     content: "\e123";
   }`;
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
 }
