@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { IconsUsageComponent } from './icons-usage.component';
 
@@ -6,12 +6,11 @@ describe('IconsUsageComponent', () => {
   let component: IconsUsageComponent;
   let fixture: ComponentFixture<IconsUsageComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ IconsUsageComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [IconsUsageComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(IconsUsageComponent);
