@@ -1,21 +1,16 @@
-import { Component, OnInit } from '@angular/core';
-import { IconsComponent } from '../icons.component';
+
+import { Component } from '@angular/core';
+import { HighlightModule } from 'ngx-highlightjs';
 
 @Component({
   selector: 'app-icons-usage',
+  standalone: true,
+  imports: [HighlightModule],
   templateUrl: './icons-usage.component.html',
   styleUrls: ['./icons-usage.component.scss']
 })
-export class IconsUsageComponent extends IconsComponent implements OnInit {
+export class IconsUsageComponent {
   public iconCheck = `<em class='icon icon-check'></em>`;
   public iconCheckPrimary = `<em class='icon icon-check primary-color'></em>`;
   public iconCheckFontSize20 = `<em class='icon icon-check font-size-xl'></em>`;
-
-  constructor() {
-    super();
-  }
-
-  ngOnInit() {
-  }
-
 }
