@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { MenuDemoComponent } from './menu-demo.component';
 
@@ -6,12 +6,11 @@ describe('MenuDemoComponent', () => {
   let component: MenuDemoComponent;
   let fixture: ComponentFixture<MenuDemoComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ MenuDemoComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [MenuDemoComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(MenuDemoComponent);
