@@ -1,15 +1,13 @@
-import { Component, OnInit } from '@angular/core';
+
+import { Component } from '@angular/core';
+import { MatMenuModule } from '@angular/material/menu';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-menu-demo',
+  standalone: true,
+  imports: [MatMenuModule, RouterLink],
   templateUrl: './menu-demo.component.html',
   styleUrls: ['./menu-demo.component.scss']
 })
-export class MenuDemoComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
+export class MenuDemoComponent {}
