@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { StyleguideComponent } from './styleguide.component';
 
@@ -6,12 +6,11 @@ describe('StyleguideComponent', () => {
   let component: StyleguideComponent;
   let fixture: ComponentFixture<StyleguideComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ StyleguideComponent ]
-    })
-    .compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [StyleguideComponent]
+    }).compileComponents();
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(StyleguideComponent);
