@@ -1,10 +1,15 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
+import { TranslateModule } from '@ngx-translate/core';
+
 import { ROUTING } from '../../consts/consts';
 
 @Component({
   selector: 'app-footer',
+  standalone: true,
+  imports: [CommonModule, TranslateModule],
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss'],
+  styleUrls: ['./footer.component.scss']
 })
 export class FooterComponent {
   public ROUTING = ROUTING;
