@@ -1,42 +1,48 @@
-import { Component, OnInit, HostListener, Input } from '@angular/core';
-import { ROUTING } from 'src/app/shared/consts/consts';
+import { CommonModule, DOCUMENT } from '@angular/common';
+import { Component, HostListener, Inject, Input } from '@angular/core';
+import { ROUTING } from '@app/shared/consts/consts';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-header',
+  standalone: true,
+  imports: [CommonModule, TranslateModule],
   templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss'],
+  styleUrls: ['./header.component.scss']
 })
-export class HeaderComponent implements OnInit {
-  @Input() scrolledTheme: boolean;
-
-  public scrolled: boolean = false;
-  public hide: boolean = false;
-  public showMenu: boolean = false;
-  public height: number = 0;
-  public ROUTING = ROUTING;
-
-  @HostListener('window:scroll', ['$event'])
-  onWindowScroll() {
-    const pos = document.documentElement.scrollTop;
+export class HeaderComponent {
+  @HostListener('window:scroll', ['$event']) public onWindowScroll(): void {
+    const pos = this.document.documentElement.scrollTop;
     const max = 5;
     if (pos <= max) {
       this.scrolled = false;
       this.hide = false;
     } else {
       this.scrolled = true;
+      const timeout = 10;
       setTimeout(() => {
         this.hide = true;
-      }, 10);
+      }, timeout);
     }
   }
 
-  constructor() {
+  @Input() public scrolledTheme: boolean;
+
+  public scrolled = false;
+  public hide = false;
+  public showMenu = false;
+  public height: number;
+
+  public readonly defaultValue = 0;
+  public readonly percent = 100;
+  public readonly ROUTING = ROUTING;
+
+  constructor(@Inject(DOCUMENT) private readonly document: Document) {
+    this.height = this.defaultValue;
     this.scrolledTheme = false;
   }
 
-  ngOnInit() {}
-
-  toggleMenu() {
-    this.height = !this.height ? 100 : 0;
+  public toggleMenu(): void {
+    this.height = this.height ? this.defaultValue : this.percent;
   }
 }
