@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { InfoboxComponent } from './infobox.component';
 
@@ -6,13 +6,11 @@ describe('InfoboxComponent', () => {
   let component: InfoboxComponent;
   let fixture: ComponentFixture<InfoboxComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ InfoboxComponent ]
-    })
-    .compileComponents();
-  }));
-
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [InfoboxComponent]
+    }).compileComponents();
+  });
   beforeEach(() => {
     fixture = TestBed.createComponent(InfoboxComponent);
     component = fixture.componentInstance;
