@@ -1,10 +1,9 @@
-import { Component} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
-    selector: 'app-spinner',
-    templateUrl: './spinner.component.html',
-    styleUrls: ['./spinner.component.scss']
+  selector: 'app-spinner',
+  standalone: true,
+  templateUrl: './spinner.component.html',
+  styleUrls: ['./spinner.component.scss']
 })
-export class spinnerComponent {
-    
-}
+export class SpinnerComponent {}
