@@ -2,5 +2,5 @@ export const ROUTES = {
   Article_2022_07_30_01: 'mariage-derval-mathilde-sebastien',
   Article_2023_03_16_01: 'comment-choisir-son-photographe-de-mariage',
   Article_2023_07_02_01: 'qu-est-ce-qu-une-seance-photo-day-after',
-  Article_2023_08_29_01: 'le-kit-de-survie-votre-guide-pour-votre-mariage',
+  Article_2023_08_29_01: 'le-kit-de-survie-votre-guide-pour-votre-mariage'
 };
